llm-pdf-qa/requirements.txt
llama-index>=0.10.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
langchain>=0.1.13
torch>=2.0.0
faiss-cpu>=1.7.4
huggingface-hub
tqdm
python-dotenv

